requirements.txt
tensorflow
keras
numpy
pillow
tkinter